requirements.txt
streamlit 
torch 
torchvision 
numpy 
opencv-python-headless 
pillow 
matplotlib